requirements.txt
PyYAML
gitpython
